Added the remaining single asset strats
